mass refactor + main + re-execution
Refactoring was done to increase the file tree clarity and accessibility, removing clutter. Additionally, a main script was added for a single-source point of execution for running each script one at a time via SLURM sbatch. While this solution provides a 'fire and forget' on the scripts, it is not optimized and will handle each script one at a time (instead of, say, running a job for each of the two assemblers). This was just done for the sake of minimalism, but can be optimized further as such.